Fixed MathTools::factorial to return long int instead of int. This fixes the Basis::rPart method and makes the MandatoryTest03 successful

+++ headers/MathTools.h
@@ -16,7 +16,7 @@
  */
 class MathTools {
     public:
-        static int factorial(int);
+        static long int factorial(int);
 
     private:
